chagned route name to /api/policy/motor/percentage-update.
