+++ controller/policyController/filterController.js
@@ -257,121 +257,6 @@ export const getAllMatchingRecords = async (req, res) => {
   }
 };
 
-// create a Entry in PayInExcelModel if payIn OD and TP same for PayOut
-export const createPercentageData = async (req, res) => {
-  try {
-    const {
-      policyType,
-      caseType,
-      productType,
-      subCategory,
-      companyName,
-      broker,
-      make,
-      model,
-      fuelType,
-      ncb,
-      vehicleAge,
-      seatingCapacity,
-      rto,
-      cc,
-      payInODPercentage,
-      payInTPPercentage,
-      payOutODPercentage,
-      payOutTPPercentage,
-      startDate,
-      endDate,
-    } = req.body;
-
-    if (
-      !policyType ||
-      !caseType ||
-      !productType ||
-      !subCategory ||
-      !companyName ||
-      !broker ||
-      !make ||
-      !model ||
-      !fuelType ||
-      !ncb ||
-      !vehicleAge ||
-      !seatingCapacity ||
-      !rto ||
-      !cc ||
-      !startDate ||
-      !endDate
-    ) {
-      return res.status(400).json({ message: "Missing required fields" });
-    }
-
-    // Parse dates
-    const start = new Date(startDate);
-    const end = new Date(endDate);
-
-    const query = {
-      policyType: policyType.toLowerCase(),
-      caseType: caseType.toLowerCase(),
-      productType: productType.toLowerCase(),
-      subCategory: subCategory.toLowerCase(),
-      companyName: companyName.toLowerCase(),
-      broker: broker.toLowerCase(),
-      make: make.toLowerCase(),
-      model: model.toLowerCase(),
-      fuelType: fuelType.toLowerCase(),
-      ncb: ncb.toLowerCase(),
-      vehicleAge: vehicleAge.toLowerCase(),
-      seatingCapacity: seatingCapacity,
-      rto: rto.toLowerCase(),
-      cc: cc,
-      startDate: start,
-      endDate: end,
-    };
-
-    const createRecord = async (Model, data) => {
-      try {
-        const newRecord = await Model.create({
-          ...query,
-          od: data.od,
-          tp: data.tp,
-          createdBy: "admin",
-          createdOn: new Date(),
-          updatedBy: null,
-          updatedOn: null,
-        });
-      } catch (err) {
-        console.error(
-          `Error creating record in ${Model.collection.name}:`,
-          err
-        );
-      }
-    };
-
-    // Check and log if payInODPercentage and payInTPPercentage are defined
-    if (payInODPercentage !== undefined && payInTPPercentage !== undefined) {
-      await createRecord(PayInExcelDataModel, {
-        od: payInODPercentage,
-        tp: payInTPPercentage,
-      });
-    }
-
-    if (payOutODPercentage !== undefined && payOutTPPercentage !== undefined) {
-      await createRecord(PayOutExcelDataModel, {
-        od: payOutODPercentage,
-        tp: payOutTPPercentage,
-      });
-    }
-
-    res.status(200).json({
-      message: "Data added successfully to the relevant models",
-      status: "Success",
-    });
-  } catch (error) {
-    res
-      .status(500)
-      .json({ message: "Error processing request", error: error.message });
-  }
-};
-
 // Update OD and TP by Date Range
 export const updateODTPByDateRange = async (req, res) => {
   const { startDate, endDate, od, tp } = req.body;
